Extract toggleFavorite to store, add addToFavorites, add route /addToken/:tokenId which will automatically add the token to the list of favorites
