+++ src/components/tokenItems/tokenItemFT.vue
@@ -278,15 +278,6 @@
       color: "red"
     }) 
   }
-
-  function toggleFavorite(tokenId: string) {
-    const newFeaturedTokens = settingsStore.featuredTokens.includes(tokenId) ?
-      settingsStore.featuredTokens.filter((id) => id !== tokenId) :
-      [...settingsStore.featuredTokens, tokenId];
-    localStorage.setItem("featuredTokens", JSON.stringify(newFeaturedTokens));
-    settingsStore.featuredTokens = newFeaturedTokens;
-    store.updateTokenList();
-  }
 </script>
 
 <template id="token-template">
@@ -333,7 +324,7 @@
             <img id="authIcon" class="icon" :src="settingsStore.darkMode? 'images/shieldLightGrey.svg' : 'images/shield.svg'">
             <span>auth transfer</span>
           </span>
-          <span @click="toggleFavorite(tokenData.tokenId)" style="float:right">
+          <span @click="store.toggleFavorite(tokenData.tokenId)" style="float:right">
             {{ settingsStore.featuredTokens.includes(tokenData.tokenId) ? "★" : "☆" }} favorite </span>
         </div>
         <div v-if="displayTokenInfo" style="margin-top: 10px;">
