+++ src/router/routes.ts
@@ -4,9 +4,14 @@ const routes: RouteRecordRaw[] = [
   {
     path: '/',
     component: () => import('pages/WalletPage.vue'),
-    props: route => ({ uri: route.query.uri })
+    props: route => ({ uri: route.query.uri, addTokenId: route.redirectedFrom?.params?.tokenId })
   },
-
+  {
+    path: '/addToken/:tokenId',
+    redirect: to => {
+      return { path: '/' }
+    }
+  },
   {
     path: '/:catchAll(.*)*',
     component: () => import('pages/ErrorNotFound.vue'),
